fix(index): preserve rebuild consistency across backends
