+++ loop-run-log.md
@@ -23,6 +23,20 @@ Append one entry per run. Prune entries older than 30 days.
 
 ```json
 {
+  "run_id": "2026-07-26T10:03:13Z",
+  "pattern": "pr-review-repair",
+  "duration_s": 720,
+  "items_found": 4,
+  "actions_taken": 4,
+  "escalations": 0,
+  "tokens_estimate": 18000,
+  "outcome": "fix-proposed",
+  "summary": "PR #1172 third Codex Draft review: force rebuilds now recreate imports and graph edges without FTS5, roll back the destructive clear if any backend cleanup fails, and retract new/updated/unchanged counters when late snapshot mutations invalidate work; snapshot regressions were consolidated into the canonical AST-cache and incremental-sync test modules. Focused coverage (240 tests), strict worktree patch coverage, full pytest (1282 passed, 7 skipped), build, Ruff, changed-file formatting, focused MyPy, test-quality audit, weak-assertion ratchet, and diff checks passed."
+}
+```
+
+```json
+{
   "run_id": "2026-07-26T09:36:38Z",
   "pattern": "pr-review-repair",
   "duration_s": 900,
